pacifico-site-layouts/backend/requirements.txt
# =============================================================================
# Pacifico Site Layouts - Backend Dependencies
# =============================================================================

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
psycopg[binary]>=3.2.0
alembic>=1.14.0

# PostGIS Support
geoalchemy2>=0.15.0

# Configuration
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# AWS
boto3>=1.35.0

# Authentication
python-jose[cryptography]>=3.3.0

# File Processing (for KML/KMZ in A-05)
fastkml>=1.0.0
shapely>=2.0.0
python-multipart>=0.0.9

# HTTP Client (for JWKS fetching and DEM in B-01)
httpx>=0.27.0

# Geospatial Processing (Phase B)
# rasterio>=1.4.0
# numpy>=2.0.0

# PDF Generation (Phase B)
# reportlab>=4.2.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
